feat: move template helper functions from react tempalte
