+++ packages/sdk/src/codegen/templates/typescript/react.ts
@@ -7,6 +7,7 @@ import { OperationType } from '@wundergraph/protobuf';
 import { template as providerTemplate } from './react.provider.template';
 import { template as reactNativeProviderTemplate } from './react.native.provider.template';
 import { template as hooksTemplate } from './react.hooks.template';
+import { hasInput, isNotInternal } from './helpers';
 
 export class TypescriptReactProvider implements Template {
 	generate(config: ResolvedWunderGraphConfig): Promise<TemplateOutputFile[]> {
@@ -80,18 +81,18 @@ export class TypescriptReactHooks implements Template {
 	}
 }
 
-export const isNotInternal = (op: GraphQLOperation): boolean => !op.Internal;
+// export const isNotInternal = (op: GraphQLOperation): boolean => !op.Internal;
 
-export const hasInput = (op: GraphQLOperation): boolean =>
-	op.VariablesSchema.properties !== undefined && Object.keys(op.VariablesSchema.properties).length !== 0;
+// export const hasInput = (op: GraphQLOperation): boolean =>
+// 	op.VariablesSchema.properties !== undefined && Object.keys(op.VariablesSchema.properties).length !== 0;
 
-export const hasInternalInput = (op: GraphQLOperation): boolean =>
-	op.InternalVariablesSchema.properties !== undefined &&
-	Object.keys(op.InternalVariablesSchema.properties).length !== 0;
+// export const hasInternalInput = (op: GraphQLOperation): boolean =>
+// 	op.InternalVariablesSchema.properties !== undefined &&
+// 	Object.keys(op.InternalVariablesSchema.properties).length !== 0;
 
-export const hasInjectedInput = (op: GraphQLOperation): boolean =>
-	op.InjectedVariablesSchema.properties !== undefined &&
-	Object.keys(op.InjectedVariablesSchema.properties).length !== 0;
+// export const hasInjectedInput = (op: GraphQLOperation): boolean =>
+// 	op.InjectedVariablesSchema.properties !== undefined &&
+// 	Object.keys(op.InjectedVariablesSchema.properties).length !== 0;
 
 const mustOperationTypeQuery = (op: GraphQLOperation) => op.OperationType === OperationType.QUERY;
 const mustLiveQuery = (op: GraphQLOperation) => op.LiveQuery !== undefined && op.LiveQuery.enable;
